+++ test/test_command_line.py
@@ -1,7 +1,0 @@
-from unittest import TestCase
-
-from seqcluster.command_line import main
-
-class TestCommand_line(TestCase):
-    def test_basic(self):
-        None